Move FreeText1-3 to a trait
